Fixed types.h

+++ src/include/types.h
@@ -10,8 +10,8 @@ typedef unsigned short uint16;
 typedef signed int int32;
 typedef unsigned int uint32;
 
-typedef signed long int64;
-typedef unsigned long uint64;
+typedef signed long long int64;
+typedef unsigned long long uint64;
 
 typedef char* string;
